move backend dependent code to AbstractAction respective classes
